backend/requirements-docker.txt
# Minimal FastAPI dependencies for Docker build
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Basic authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0

# Database essentials
sqlalchemy[asyncio]==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI APIs (lightweight)
openai==1.3.7
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
python-magic==0.4.27
Pillow==10.1.0

# Caching and Performance
redis==5.0.1
psutil==5.9.6

# Distributed Tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-jaeger-thrift==1.21.0

# Production Essentials
alembic==1.13.1
bcrypt==4.1.2
cryptography==41.0.8
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
structlog==23.2.0
gunicorn==21.2.0

# Advanced Features
celery==5.3.4
flower==2.0.1
validators==0.22.0
email-validator==2.1.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0